# swagger.yml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: http://localhost:8000/api/v1
info:
  description: This is a BT API document
  version: 1.0.1
  title: BT API document
  contact:
    email: baotrandinh100@gmail.com
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
paths:
  /auth/login:
    post:
      security: []
      tags:
        - Auth
      description: |
        User Login by Email and Password
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCredentials"
        description: email and password
      responses:
        "200":
          description: login successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCredentialResponse"
        "400":
          description: invalid email or password format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: incorrect email or password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/signup:
    post:
      tags:
        - Auth
      description: User sign up an local account
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSignUp"
      responses:
        "200":
          description: Server will send an email to user to active this account then it sets a cookie with the value {"active_account":false, expired:60s}
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouceSuccess"
        "400":
          description: errors of the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/seller/register:
    get:
      tags:
        - Auth
      security:
        - bearerAuth: []
      description: send request to sign up seller
      responses:
        "200":
          description: Server will send an email to user to active this account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouceSuccess"
        "404":
          description: errors of the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/token/check:
    get:
      tags:
        - Auth
      description: check token expire or exists
      responses:
        "200":
          description: Server will send a notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouceSuccess"
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: errors of the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/seller/create:
    post:
      tags:
        - Auth
      description: create a new seller
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserSignUp"
      responses:
        "200":
          description: Server will send a notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouceSuccess"
        "400":
          description: bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: errors of the server
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/password/email:
    post:
      tags:
        - Auth
      description: Send mail to reset password 
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Server will send Link reset password to your email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouceSuccess"
        "400":
            description: Invalid Email Format/Email is not Registered
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"
        "500":
            description: Internal Server Error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"  
  /auth/password/reset:
    post:
      tags:
        - Auth
      description: Token matched then server will send for you verify token to reset password
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "200":
          description: Server will send verify code to you, then you get it  to reset password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouceSuccess"
        "400":
            description: bad request
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"
        "500":
            description: Internal Server Error
            content:
              application/json:
                schema:
                  $ref: "#/components/schemas/ErrorResponse"
  /auth/login/refresh:
    post:
      tags:
        - Auth
      description: Refresh token when access token is expired
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: return a new payload + access_token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCredentials"
        "400":
          description: error token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/oauth/google:
    get:
      tags:
        - Auth
      description: Login by google account
      responses:
        "200":
          description: redirect user for the web page
        "400":
          description: errors are from google api
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT  
  schemas:
...
    UserCredentials:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          example: baotrandinh100@gmail.com
        password:
          type: string
          example: abc#1234
    UserSignUp:
      type: object
      required:
        - email
        - password
        - name
        - gender
        - birthDay
      properties:
        email:
          type: string
          example: baotrandinh100@gmail.com
        password:
          type: string
          description: Password must be 8-16 characters and contain both numbers and letters characters
          example: 12345678b
        name:
          type: string
        gender:
          type: string
          description: it kind of a checkbok
          example: "male | female "
        bithDay:
          type: string
    RefreshTokenRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
    UserCredentialResponse:
      type: object
      required:
        - data
      properties:
        status:
          type: number
        message:
          type: string
        data:
          type: object
          required:
            - name
            - avatar
            - role
            - gender
            - typeLogin
            - accessToken
            - refreshToken
          properties:
            userId:
              type: string
            name:
              type: string
            nickName:
              type: string
            role:
              type: string
              example: user | seller
            avatar:
              type: string
            meta:
              type: object
              properties:
                totalBuy:
                  type: integer
                totalCancel:
                  type: integer
            typeLogin:
              type: string
              example: local | google
            seller:
              $ref: "#/components/schemas/SellerDetail"
            special:
              type: array
              items:
                type: object
...
    ErrorResponse:
      description: Error responses are sent when an error (e.g. unauthorized, bad request) occurred.
      type: object
      properties:
        status:
          description: the status of response.
          type: integer
          example: 400
        errors:
          type: object
          properties:
            message:
              type: string
              example: Email is existed
...
    SellerDetail:
      description: Seller detail information
      type: object
      properties:
        _id:
          type: string
          example: 63312b2a82dc64f45360ff5a
        info:
          type: object
          properties:
            name:
              type: string
            address:
              type: array
              items:
                type: string
            phone: 
              type: string
        logo:
          type: object
          properties:
            fileLink:
              type: string
            fileId:
              type: string
        slogan:
          type: string
          example: you can do it!
        type:
          type: string
          example: normal | mall | global
        meta:
          $ref: "#/components/schemas/MetaSeller"
        isDisable:
          description: to implement lock the account of seller
          type: boolean
          default: false
    MetaSeller:
      description: some information that it relate the shop of seller
      type: object
      properties:
        totalSole:
          type: number
          default: 0
        totalProduct:
          type: number
          default: 0
        totalEvaluation:
          type: number
          default: 0
        ranking: 
          type: number
          default: 0
        title:
          type: string
          example: the best seller
...
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          example: baotrandinh100@gmail.com
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
      description: Link got from your email
      properties:
        token:
          type: string
        password:
          type: string
          description: Password must contain at least 8 characters, including letters, numbers and special characters.
...
    AnnouceSuccess:
      type: object
      properties:
        status:
          type: integer
          example: 200
        message:
          type: string
          example: "register is success, please check your email to active account"
        data:
          type: object
...